Clean-up spot.py from duplication
