+++ bd_spot_wrapper/spot_wrapper/spot.py
@@ -26,7 +26,7 @@ import numpy as np
 import quaternion
 import rospy
 import sophus as sp
-from aria_data_utils.conversions import sophus_to_ros_pose
+from aria_data_utils.conversions import sophus_to_ros_pose, generate_TransformStamped_a_T_b_from_SE3Pose
 from bosdyn import geometry
 from bosdyn.api import (
     arm_command_pb2,
@@ -59,14 +59,8 @@ from bosdyn.client.robot_command import (
 )
 from bosdyn.client.robot_state import RobotStateClient
 from bosdyn.util import seconds_to_duration
-
-############################################################################
-# TODO: Move to transform_utils.py
-from geometry_msgs.msg import Pose, PoseStamped, TransformStamped
 from google.protobuf import wrappers_pb2
-from scipy.spatial.transform import Rotation as R
-
-############################################################################
+from geometry_msgs.msg import Pose
 
 # Get Spot password and IP address
 env_err_msg = (
@@ -132,146 +126,6 @@ SHOULD_ROTATE = [
     SpotCamIds.HAND_DEPTH,
     SpotCamIds.HAND,
 ]
-
-
-# TODO: Move to transform_utils.py
-def generate_TransformStamped_a_T_b_from_SE3Pose(
-    a_Tform_b: SE3Pose, parent_frame: str, child_frame: str
-) -> TransformStamped:
-    """Generate the transform from spotWorld to spot frame
-
-    Returns:
-        Transform: Transform from spotWorld to spot
-    """
-    transform_a_T_b = TransformStamped()
-
-    transform_a_T_b.header.stamp = rospy.Time.now()
-    transform_a_T_b.header.frame_id = f"/{parent_frame}"
-    transform_a_T_b.child_frame_id = f"/{child_frame}"
-
-    trans = a_Tform_b.position
-    transform_a_T_b.transform.translation.x = float(trans.x)
-    transform_a_T_b.transform.translation.y = float(trans.y)
-    transform_a_T_b.transform.translation.z = float(trans.z)
-
-    quat = a_Tform_b.rotation
-    transform_a_T_b.transform.rotation.x = float(quat.x)
-    transform_a_T_b.transform.rotation.y = float(quat.y)
-    transform_a_T_b.transform.rotation.z = float(quat.z)
-    transform_a_T_b.transform.rotation.w = float(quat.w)
-    return transform_a_T_b
-
-
-# TODO: Move to transform_utils.py
-def generate_TransformStamped_a_T_b_from_spSE3(
-    a_Tform_b: sp.SE3, parent_frame: str, child_frame: str
-) -> TransformStamped:
-    """Generate the transform from spotWorld to spot frame
-
-    Returns:
-        Transform: Transform from spotWorld to spot
-    """
-    transform_a_T_b = TransformStamped()
-
-    transform_a_T_b.header.stamp = rospy.Time.now()
-    transform_a_T_b.header.frame_id = f"/{parent_frame}"
-    transform_a_T_b.child_frame_id = f"/{child_frame}"
-
-    trans = a_Tform_b.translation()
-    transform_a_T_b.transform.translation.x = float(trans[0])
-    transform_a_T_b.transform.translation.y = float(trans[1])
-    transform_a_T_b.transform.translation.z = float(trans[2])
-
-    quat = R.from_matrix(a_Tform_b.rotationMatrix()).as_quat()
-    transform_a_T_b.transform.rotation.x = float(quat[0])
-    transform_a_T_b.transform.rotation.y = float(quat[1])
-    transform_a_T_b.transform.rotation.z = float(quat[2])
-    transform_a_T_b.transform.rotation.w = float(quat[3])
-    return transform_a_T_b
-
-
-# TODO: Move to transform_utils.py
-def generate_PoseStamped_a_T_b_from_spSE3(
-    a_Tform_b: sp.SE3, parent_frame: str
-) -> PoseStamped:
-    """Generate the PoseStamped for frame b w.r.t frame a
-
-    Returns:
-        Transform: Posestamp for frame b w.r.t frame a
-    """
-    pose_a_T_b = PoseStamped()
-
-    # pose_a_T_b.header.seq = 0 # TODO: Not sure if this is needed
-    pose_a_T_b.header.stamp = rospy.Time.now()
-    pose_a_T_b.header.frame_id = f"{parent_frame}"
-
-    trans = a_Tform_b.translation()
-    pose_a_T_b.pose.position.x = float(trans[0])
-    pose_a_T_b.pose.position.y = float(trans[1])
-    pose_a_T_b.pose.position.z = float(trans[2])
-
-    quat = R.from_matrix(a_Tform_b.rotationMatrix()).as_quat()
-    pose_a_T_b.pose.orientation.x = float(quat[0])
-    pose_a_T_b.pose.orientation.y = float(quat[1])
-    pose_a_T_b.pose.orientation.z = float(quat[2])
-    pose_a_T_b.pose.orientation.w = float(quat[3])
-    return pose_a_T_b
-
-
-# TODO: Move to transform_utils.py
-def generate_spSE3_a_T_b_from_TransformStamped(a_Tform_b: TransformStamped) -> sp.SE3:
-    """Generate the transform from spotWorld to spot frame
-
-    Returns:
-        Transform: Transform from spotWorld to spot
-    """
-
-    trans = np.array([0.0, 0.0, 0.0])
-    trans[0] = a_Tform_b.transform.translation.x
-    trans[1] = a_Tform_b.transform.translation.y
-    trans[2] = a_Tform_b.transform.translation.z
-
-    quat = np.array([0.0, 0.0, 0.0, 1.0])
-    quat[0] = a_Tform_b.transform.rotation.x
-    quat[1] = a_Tform_b.transform.rotation.y
-    quat[2] = a_Tform_b.transform.rotation.z
-    quat[3] = a_Tform_b.transform.rotation.w
-    return sp.SE3(
-        R.from_quat(quat).as_matrix(), trans
-    )  # R.from_quat() takes a quaternion in the form of [x, y, z, w]
-
-
-# TODO: Move to transform_utils.py
-def generate_spSE3_a_T_b_from_PoseStamped(a_Tform_b: PoseStamped) -> sp.SE3:
-    """Generate the transform from spotWorld to spot frame
-
-    Returns:
-        Transform: Transform from spotWorld to spot
-    """
-
-    trans = np.array([0.0, 0.0, 0.0])
-    trans[0] = a_Tform_b.pose.position.x
-    trans[1] = a_Tform_b.pose.position.y
-    trans[2] = a_Tform_b.pose.position.z
-
-    quat = np.array([0.0, 0.0, 0.0, 1.0])
-    quat[0] = a_Tform_b.pose.orientation.x
-    quat[1] = a_Tform_b.pose.orientation.y
-    quat[2] = a_Tform_b.pose.orientation.z
-    quat[3] = a_Tform_b.pose.orientation.w
-    return sp.SE3(
-        R.from_quat(quat).as_matrix(), trans
-    )  # R.from_quat() takes a quaternion in the form of [x, y, z, w]
-
-
-# TODO: Move to transform_utils.py
-def xyt2sophus(xyt: np.ndarray) -> sp.SE3:
-    """
-    Converts SE2 coordinates (x, y, rz) to an sophus SE3 pose object.
-    """
-    x = np.array([xyt[0], xyt[1], 0.0])
-    r_mat = sp.SO3.exp([0.0, 0.0, xyt[2]]).matrix()
-    return sp.SE3(r_mat, x)
 
 
 class Spot:
